requirements.txt
black==23.3.0
fastparquet==2023.4.0
kafka-python==2.0.2
pandas==2.0.1
pylint==2.17.4